Cyton API

These are the types and methods that implement the Cyton framework.

See this guide for an overview: cyton-dynamics.github.io/

Cyton.CellPopulation — Type

CellPopulation

The population of cells with convenince constructors. Ordinarily, this will be constructed by the framework.

Cyton.step — Method

step(::FateTimer, ::Time, ::Duration)::Union{CellEvent, Nothing}

Step the FateTimer forward by one time increment. This should be implemented for each concrete FateTimer

Optionally return a CellEvent

Cyton.inherit — Function

inherit(::FateTimer, ::Time)::FateTimer

Called when a cell divides, this allows daughters to inherit timers, create new timers or something in between.

Note: It may be tempting to simple return the parent timer if the cell inherits the timers but be aware the stepper will called for that timer for each cell that references it.

Cyton.modelTime — Method

modelTime(model::CellPopulation)::Time

Returns the current model time

Cyton.modelTimeStep — Method

modelTimeStep(model::CellPopulation)::Duration

Returns the current model time step.

Cyton.cellCount — Method

cellCount(model::CellPopulation)::Int

Return the number of cells in the population

Cyton.createPopulation — Method

createPopulation(nCells::Int, cellFactory::Function; eventCallbacks::Vector{Function}=Function[])::CellPopulation

Create a population of cells: nCells: size of starting populations cellFactory: A function that returns constructs a new cell eventCallbacks: Function that are called when events occurs

Division and Death are two predefined events

Cyton.step — Method

step(model::CellPopulation, stimulus::T) where T<:Stimulus

Step the population forward in time by one time step, with a single stimulus.

Cyton.step — Method

step(model::CellPopulation, stimuli::Vector{T}=Vector{Stimulus}()) where T<:Stimulus

Step the population forward in time by one time step, with optional stimuli

Cyton.step — Method

step(agent::CellAgent, model::CellPopulation, stimuli::Vector{T}) where T<:Stimulus

Step a cell forward in time by one time step.