Cyton API

These are the types and methods that implement the Cyton framework.

See this guide for an overview: cyton-dynamics.github.io/

Types and constructors

Events

Cyton.CellEvent — Type

CellEvent

Top level type for modelling cellular events. Death and Division are handled internally but users should add additional events to model additional behaviour.

Cyton.Division — Type

Division

A timer (or something else) has triggered cell division

Cyton.Death — Type

Death

A timer (or something else) has triggered a cell death event

Cyton.Time — Type

Time

A type alias for time-like quantities. Used to make function signatures more readable

Cyton.Duration — Type

Duration

A type alias for duration-like quantities. Used to make function signatures more readable

Cells and populations

Cyton.CellPopulation — Type

CellPopulation

The population of cells with convenince constructors. Ordinarily, this will be constructed by the framework.

Cyton.CellType — Type

CellType

This is a type (in the software sense) that a cell can carry to distinguish different categories of cell, e.g. WildType, Knockout, etc

Cyton.GenericCell — Type

GenericCell

A basic cell type, which is used if no more specific type is provided by the modeller. This can be used when all cells in the model are the same.

Cyton.AbstractCell — Type

AbstractCell

The base cell type for objects that carry individual cell state

Cyton.Cell — Type

Cell{T} <: AbstractCell{T}

The main user facing Cell data type. It encapsulates another type, T, which can model concepts such as phenotype or genotype

Cyton.Cell — Method

Cell(birth::Time)::Cell{GenericCell}

Constructor for a generic cell.

Cyton.Cell — Method

Cell(birth::Time, cellType::T) where T <: CellType

Constructor for a cell of type T

Cyton.Cell — Method

Cell(birth::Time, divisionCount::Int64)::Cell{GenericCell}

Constructor for a daughter cell (i.e. a cell with a division count)

Stimuli

Cyton.Stimulus — Type

Stimulus

Top level type for modelling cell stimuli

API methods and functions

These are the methods than need to be implemented to build a model.

Cyton.stimulate — Function

stimulute(::Cell, ::Stimulus, time::Time, Δt::Duration)

This function is called when a stimulus is applied to a cell. This function is overriden by the modeller to implement their required behaviour.

Cyton.step — Method

step(::FateTimer, ::Time, ::Duration)::Union{CellEvent, Nothing}

Step the FateTimer forward by one time increment. This should be implemented for each concrete FateTimer

Optionally return a CellEvent

Cyton.inherit — Function

inherit(::FateTimer, ::Time)::FateTimer

Called when a cell divides, this allows daughters to inherit timers, create new timers or something in between.

Note: It may be tempting to simple return the parent timer if the cell inherits the timers but be aware the stepper will called for that timer for each cell that references it.

Utility methods

Cyton.addTimer — Function

addTimer

Add a FateTimer to a cell.

Cyton.age — Function

age

Return the age of the cell

Cyton.addObserver — Function

addObserver

Adds a callback function, observer(event::CellEvent, cell::Cell, time::Time), to a cell. This function is called when Cell triggers a CellEvent.

Cyton.modelTime — Method

modelTime(model::CellPopulation)::Time

Returns the current model time

Cyton.modelTimeStep — Method

modelTimeStep(model::CellPopulation)::Duration

Returns the current model time step.

Cyton.cellCount — Method

cellCount(model::CellPopulation)::Int

Return the number of cells in the population

Cyton.cohortCount — Method

cohortCount(model::CellPopulation)::Int

Return the current cohort count, cell count normalised by generation number.

Cyton.createPopulation — Method

createPopulation(nCells::Int, cellFactory::Function; eventCallbacks::Vector{Function}=Function[])::CellPopulation

Create a population of cells: nCells: size of starting populations cellFactory: A function that returns constructs a new cell eventCallbacks: Function that are called when events occurs

Division and Death are two predefined events

Cyton.step — Method

step(model::CellPopulation, stimulus::T) where T<:Stimulus

Step the population forward in time by one time step, with a single stimulus. This is called in the modeller's simulation loop.

Cyton.step — Method

step(model::CellPopulation, stimuli::Vector{T}=Vector{Stimulus}()) where T<:Stimulus

Step the population forward in time by one time step, with optional stimuli

Probability distributions

Cyton.DistributionParmSet — Type

DistributionParmSet

A type holding parameters for a distribution.

Cyton.PoissonParms — Type

PoissonParms

A Poisson distribution

Cyton.FixedDistributionParms — Type

FixedDistributionParms

This distribution draws the same value every time.

Cyton.draw — Method

draw(d::DistributionParmSet)

Return a random read number from the distribution.

Cyton.usefulMax — Function

usefulMax

Returns a hint for plotting functions.

Cyton.plotPdf — Function

plotPdf

Plot the probability density function for this distribution.

Cyton.NormalParms — Type

NormalParms

A normal distribution.

Cyton.LogNormalParms — Type

LogNormalParms

A log normal distribution parameterised as $P(x; \mu, \sigma) = \frac{1}{x \sigma \sqrt{2 \pi}} \exp(-\frac{(\ln x - \mu)^2}{2 \sigma^2})$